Fix: product-service connection to gateaway-service

+++ product-service/models/product.go
@@ -1,10 +1,10 @@
 package models
 
 type Product struct {
-	Id      int     `json:"id" gorm:"primary_key"`
-	Title   string  `json:"title"`
-	Content string  `json:"content"`
-	Image   string  `json:"image"`
-	Price   float64 `json:"Price"`
-	Stock	int		`json:"Stock"`
+	Id      int    `json:"id" gorm:"primary_key"`
+	Title   string `json:"title"`
+	Content string `json:"content"`
+	Image   string `json:"image"`
+	Price   int    `json:"Price"`
+	Stock   int    `json:"Stock"`
 }
